Adapt pg_pathman for PG 11.
I kinda lost interest to exorcise a couple of tests further in attempts to make
them pass on all supported versions and just added copies. These are
 * pathman_expressions now differs because planner converts ROW(Const, Const) to
   just Const of record type.
 * Same with pathman_rebuild_updates.

I have removed inclusion of partition_filter.h in pg_compat.h in 9.5 as it
created circular dependency hell. I think it is not worthwhile to fight with it
since the only thing actually needed was error message, which is used in this
single place.

Small typo fix in partitioning_test.py: con2.begin instead of con1.begin.

Finally, run python tests with --failfast and --verbose options.